Create mutableIn struct
The mutableIn struct satisfies both the reader/writer interfaces. This
would allow users to get input from the struct's buffer and inject data
into it.

This functionality is helpful because as the project expands and terminal
input emulation is added, mutableIn could be used as the primary input
reader over os.Stdin when a program needs to write and visualize input in
a CLI for a user.