Panic if isRunning is false
I have updated the Read() and Write() methods to panic if isRunning is
false. This is necessary because mutableIn seeks to mimic the standard
input stream. In the future this will be done by visually emulating input
while the terminal is in a raw state. If the terminal is not in a raw
state, with the future emulation and other planned functionalities, then
there wouldn't be any point to read and write to its buffer. So its
enforced by ensuring the user can't use those functions unless the Init()
function is properly used first.